+++ src/Modules/funcs.js
@@ -1,0 +1,20 @@
+export const getOrdinalSuffix = (day) => {
+    if (day > 3 && day < 21) return "th";
+    switch (day % 10) {
+      case 1:
+        return "st";
+      case 2:
+        return "nd";
+      case 3:
+        return "rd";
+      default:
+        return "th";
+    }
+  };
+
+const today = new Date();
+const day = today.getDate();
+const month = today.toLocaleString("default", { month: "long" });
+const year = today.getFullYear();
+
+export const formattedDate = `${day}${getOrdinalSuffix(day)}, ${month} ${year}`;
