+++ src/output.css
@@ -725,28 +725,12 @@ video {
   margin-top: 2rem;
 }
 
-.mt-\[\] {
-  margin-top: ;
-}
-
-.mt-auto {
-  margin-top: auto;
-}
-
 .mt-\[64px\] {
   margin-top: 64px;
 }
 
-.ml-\[\] {
-  margin-left: ;
-}
-
-.ml-\[64px\] {
-  margin-left: 64px;
-}
-
-.ml-\[248px\] {
-  margin-left: 248px;
+.mt-auto {
+  margin-top: auto;
 }
 
 .block {
@@ -1008,6 +992,22 @@ video {
 
 .transform {
   transform: translate(var(--tw-translate-x), var(--tw-translate-y)) rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y)) scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
+}
+
+@keyframes bounce {
+  0%, 100% {
+    transform: translateY(-25%);
+    animation-timing-function: cubic-bezier(0.8,0,1,1);
+  }
+
+  50% {
+    transform: none;
+    animation-timing-function: cubic-bezier(0,0,0.2,1);
+  }
+}
+
+.animate-bounce {
+  animation: bounce 1s infinite;
 }
 
 @keyframes pulse {
@@ -1409,6 +1409,11 @@ video {
   background-color: rgb(34 85 34 / var(--tw-bg-opacity));
 }
 
+.bg-pink-500 {
+  --tw-bg-opacity: 1;
+  background-color: rgb(236 72 153 / var(--tw-bg-opacity));
+}
+
 .bg-white {
   --tw-bg-opacity: 1;
   background-color: rgb(255 255 255 / var(--tw-bg-opacity));
@@ -1559,10 +1564,6 @@ video {
 
 .pt-4 {
   padding-top: 1rem;
-}
-
-.pl-\[248px\] {
-  padding-left: 248px;
 }
 
 .text-left {
@@ -1937,11 +1938,15 @@ video {
   --tw-bg-opacity: 1;
   background-color: rgb(255 255 255 / var(--tw-bg-opacity));
   --tw-text-opacity: 1;
-  color: rgb(0 0 0 / var(--tw-text-opacity));
+  color: rgb(34 85 34 / var(--tw-text-opacity));
   transition-property: all;
   transition-duration: 200ms;
   transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
 }
+
+/* .activeImg {
+  @apply color
+} */
 
 .activeText::before {
   padding: 10px;
@@ -2363,10 +2368,6 @@ video {
 @media (min-width: 1024px) {
   .lg\:sticky {
     position: sticky;
-  }
-
-  .lg\:ml-\[248px\] {
-    margin-left: 248px;
   }
 
   .lg\:ml-64 {
