--- src/pages/dashboard/NotificationCard.jsx
@@ -1,15 +1,0 @@
-import { useParams } from 'react-router-dom';
-
-
-const NotificationCard = () => {
-    const { id } = useParams()
-    console.log(id)
-
-    return (
-        <div>
-            {id}
-        </div>
-    )
-}
-
-export default NotificationCard
