--- src/pages/dashboard/Userprofile.jsx
@@ -1,22 +1,0 @@
-import Profile from "../Personnel/pages/Profile";
-
-const Userprofile = () => {
-  // const currentUserId = "AgboolaSSRL001";
-  // const user = {
-  //   name: "Agboola Oluwatofunmi Bolutife",
-  //   id: "AgboolaSSRL001",
-  //   stack: "Software",
-  //   niche: "Web Development",
-  //   role: "intern",
-  //   dateJoined: "April 20th, 2024",
-  //   bio: "An upcoming web developer at Smart System Research Laboratory",
-  //   dob: "May 1st, 2003",
-  //   location: "Southgate",
-  //   contact: "09057299991",
-  //   email: "oluwatofunmiagboola1@gmail.com",
-  // };
-
-  return <Profile />;
-};
-
-export default Userprofile;
